Added provider v3.0.0 package; code refactor; updated gitignore
