+++ .gitignore
@@ -15,6 +15,10 @@
 *.iws
 .idea/
 
+android/**/.project
+android/**/.settings
+android/**/.classpath
+
 # The .vscode folder contains launch configuration and tasks you configure in
 # VS Code which you may wish to be included in version control, so this line
 # is commented out by default.
